+++ api/tests/acceptance/application/organizations/organization-controller_test.js
@@ -2,8 +2,6 @@ import lodash from 'lodash';
 
 import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
 import { ORGANIZATION_FEATURE } from '../../../../src/shared/domain/constants.js';
-import { Membership } from '../../../../src/shared/domain/models/Membership.js';
-import { OrganizationInvitation } from '../../../../src/team/domain/models/OrganizationInvitation.js';
 import {
   createServer,
   databaseBuilder,
@@ -12,7 +10,6 @@ import {
   insertMultipleSendingFeatureForNewOrganization,
   insertUserWithRoleSuperAdmin,
   knex,
-  sinon,
 } from '../../../test-helper.js';
 
 const { map: _map } = lodash;
@@ -397,76 +394,6 @@ describe('Acceptance | Application | organization-controller', function () {
         expect(response.statusCode).to.equal(200);
         expect(response.result.meta).to.deep.equal(expectedMetaData);
         expect(response.result.data).to.have.lengthOf(0);
-      });
-    });
-  });
-
-  describe('PATCH /api/organizations/{id}/resend-invitation', function () {
-    let clock;
-    const now = new Date('2022-12-25');
-
-    beforeEach(function () {
-      clock = sinon.useFakeTimers({
-        now,
-        toFake: ['Date'],
-      });
-    });
-
-    afterEach(async function () {
-      clock.restore();
-    });
-
-    it('should return the matching organization invitation as JSON API', async function () {
-      // given
-      const adminUserId = databaseBuilder.factory.buildUser().id;
-      const organizationId = databaseBuilder.factory.buildOrganization().id;
-      databaseBuilder.factory.buildMembership({
-        userId: adminUserId,
-        organizationId,
-        organizationRole: Membership.roles.ADMIN,
-      });
-
-      const email = 'anna.tole@example.net';
-      const userToReInvite = databaseBuilder.factory.buildUser({ email });
-      const existingOrganizationInvitationId = databaseBuilder.factory.buildOrganizationInvitation({
-        organizationId,
-        email,
-        updatedAt: new Date('2022-12-12'),
-      }).id;
-
-      await databaseBuilder.commit();
-
-      // when
-      const response = await server.inject({
-        method: 'PATCH',
-        url: `/api/organizations/${organizationId}/resend-invitation`,
-        headers: generateAuthenticatedUserRequestHeaders({ userId: adminUserId }),
-        payload: {
-          data: {
-            type: 'organization-invitations',
-            attributes: {
-              email: 'annA.tole@example.net',
-            },
-          },
-        },
-      });
-
-      // then
-      expect(response.statusCode).to.equal(200);
-      expect(response.result).to.deep.equal({
-        data: {
-          id: `${existingOrganizationInvitationId}`,
-          type: 'organization-invitations',
-          attributes: {
-            'organization-id': organizationId,
-            'organization-name': undefined,
-            email: userToReInvite.email,
-            status: OrganizationInvitation.StatusType.PENDING,
-            role: null,
-            lang: 'fr',
-            'updated-at': now,
-          },
-        },
       });
     });
   });
